requirements.txt
Django
Pillow
python-decouple==3.8
djangorestframework
requests
django-cors-headers
django-rest-passwordreset